move status bar context to a "store"
