+++ apps/web/core/state/status-bar-store/index.ts
@@ -1,0 +1,1 @@
+export { StatusBarContextProvider, useStatusBar } from './status-bar-store';
