+++ apps/web/partials/review/flow-bar.tsx
@@ -10,6 +10,7 @@ import { useActionsStore } from '~/core/hooks/use-actions-store';
 import { useToast } from '~/core/hooks/use-toast';
 import { useDiff } from '~/core/state/diff-store/diff-store';
 import { useEditable } from '~/core/state/editable-store/editable-store';
+import { useStatusBar } from '~/core/state/status-bar-store';
 import { ReviewState } from '~/core/types';
 import { Action } from '~/core/utils/action';
 
@@ -87,51 +88,6 @@ export const FlowBar = () => {
     </AnimatePresence>
   );
 };
-
-interface StatusBarState {
-  reviewState: ReviewState;
-  error: string | null;
-}
-
-type StatusBarActions =
-  | {
-      type: 'SET_REVIEW_STATE';
-      payload: ReviewState;
-    }
-  | { type: 'ERROR'; payload: string | null };
-
-const statusBarReducer = (state: StatusBarState, action: StatusBarActions): StatusBarState => {
-  switch (action.type) {
-    case 'SET_REVIEW_STATE':
-      return { reviewState: action.payload, error: null };
-    case 'ERROR':
-      return { reviewState: 'publish-error', error: action.payload };
-  }
-};
-
-const StatusBarContext = React.createContext<{
-  state: StatusBarState;
-  dispatch: React.Dispatch<StatusBarActions>;
-} | null>(null);
-
-export const StatusBarContextProvider = ({ children }: { children: React.ReactNode }) => {
-  const [state, dispatch] = React.useReducer(statusBarReducer, {
-    reviewState: 'idle',
-    error: null,
-  });
-
-  return <StatusBarContext.Provider value={{ state, dispatch }}>{children}</StatusBarContext.Provider>;
-};
-
-export function useStatusBar() {
-  const context = React.useContext(StatusBarContext);
-
-  if (!context) {
-    throw new Error('useStatusBar must be used within a StatusBarContextProvider');
-  }
-
-  return context;
-}
 
 const StatusBar = () => {
   const { state, dispatch } = useStatusBar();
